Make sure GitHub Pages `index.html` works
The GitHub Pages setup is now use `/docs`, so it will try to load
`/docs/index.html`.  That file must reference the rest of the GitHub
pages from `/docs/gh-pages/` using relative references.

We also remove the unnecessary `./` references and remove the
unnecessary `package-lock.json` which updates the security scanners.