Release Notes for `v0.71.0`
This new file, `RELEASE_NOTES.md`, is being added to the `/doc` sub-
directory so that we can code-review the contents of the release notes
like we would any other code change.

The contents of the file will be used for the contents of the
[Alpha](https://github.com/distributed-system-analysis/pbench/releases/edit/v0.71.0-alpha.0),
Beta, and final releases pages on GitHub.

Added items from issue #1921.