b0.71 backport: Delete the spurious pbench3-devel RPM dependency
Fixes #2798

This is a backport of #2799 to b0.71.

AFAICT, the python3-devel dependency (which only exists for RHEL7) is
spurious and in fact harmful: there seems to be no python3-devel
package for RHEL7.9, so installation fails with an unsatisfied
dependency, whereas an RPM built without it installs without
problems.

Clean up the unnecessary gcc dependency as well.