Allow a list of ?metadata tags for simplicity (#2790)
PBENCH-637

This is a minor improvement to "ease of use" allowing multiple metadata
keys on a single `metadata` query parameter. It essentially allows a
`ParamType.LIST` to be a comma-separated string (or a list of comma-separated
strings) rather than a JSON list. It also refactors the two places that processed
query parameters into an `ApiBase` helper.