Fix "index map" APIs to use dataset name
PBENCH-644

To align with the service layer of the old dashboard, the search,
mappings, and contents APIs were designed to identify datasets using
the Elasticsearch run document MD5 (run_id or run.id) rather than the
more straightforward model of relying on the dataset name.

As we approach the verge of using TOC data in the new dashboard, fixing
this seems timely.