Unwrap lines to format properly in release notes
The target location of these changes is the GitHub release notes page
for the `v0.71.0-alpha.0` release.  Unfortunately, the way formatting
is handled in a normal GitHub markdown file and how it is handled in the
release note context is different.

Unwrapping the lines so that one paragraph is one line allows us to post
the contents of the release notes directly into the GitHub release notes
page without having to reformat to make it look nice.