Fix GET /datasets/list authorization
PBENCH-648

When I deployed my new server, I found that doing an unauthorized
/datasets/list query was failing with 500. This is because the
`UnauthorizedAccess` exception is handled only in `_dispatch` where we
authorized automatically based on a JSON body payload.

My original "vision" had been to make both `UnauthorizedAccess` and
`SchemaError` extend `APIAbort` to simplify the logic, but I lost track
of that during the review. This PR fixes that, and streamlines the error
logic a bit.