Fix pbench-fio local fio-server execution
This is a back-port of commit ea18664d0 (PR #2816) from `main`.

With the separation of local vs. remote hosts handling in pbench-fio
(0dabb2392cacb46103fed7757426bc49319b18f5) the local execution was
modified without double quotes, which results in variables not being
filled.

Similarly the path that used to be set by the ssh command is not set so
the client-result.* files are being placed into the current directory.
Let's explicitly set the output path to the usual location.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>
Co-authored-by: Peter Portante <peter.portante@redhat.com>