Move towards embedded URL parameters: metadata
PBENCH-669

We started out with all parameters in a JSON request payload to `POST`
operations following Elasticsearch conventions. While some new APIs have
used both query parameters and embedded Flask URL parameters, there are
older APIs that should be retrofitted to that model.

In this PR, I'm addressing `/datasets/metadata` (which seems like a good
place to start as the Overview page is on the cusp of using it).