Add UUIDs to command line of running Tool Meisters
In order to facilitate a "big-hammer" approach to cleaning up wayward
Tool Meister sub-systems, we add UUIDs to the command line invocations
of the remote Tool Meister processes.  We don't need to do that for the
local Redis server, Tool Data Sink, and Tool Meister process because we
have their PID files already.

We do take this opportunity, though, to verify the UUID given on the
command line against the running parameters stored in Redis.