Remove use of `benchmark_bin` environment variable
Both `pbench-fio` and `pbench-uperf` no longer allow for the benchmark
binary to be overriden by an environment variable. This was an out-dated
way for the unit tests to mock the respective benchmark behavior.

Instead, mock commands for `fio` and `uperf` are provided on the `PATH`.

For the CLI tests, those mocks are removed so that we can verify that
help and usage text is emitted before the checks for the particular
command existence (which is shown by the failing `test-CL`).

Note that this change exposes the use of the `which` command, which
(pun intended) is now added to the CI container, the development
container, and as an explicit RPM dependency.