Add local and remote pre-check for linpack
We add a local and remote pre-check function for linpack.

We now wait at most 60 seconds for the linpack process to start, and
then we wait for the PID to exit instead of a file to exist.  That way
if a file is never written that we expect, but the PID has died, we'll
not wait forever.

Further, we move the "version" check (of sorts) to the linpack driver
script to allow for the remote check to work without having the code in
two places.