Ensure use of `localhost` when stopping Redis
This is a back-port of commit c84c6ed32 (PR #2863) from `main`.

When `pbench-tool-meister-stop` is invoked where the Redis server was
created locally by `pbench-tool-meister-start`, we always want to use
the `localhost` IP address to talk to it.

Also added unit tests for the `tool_meister_stop.py` module's
`RedisServer` class, and corrected the `pbench-tool-meister-stop` CLI
param default value for the `--redis-server` switch.

Fixes issue #2861 [1].

[1] https://github.com/distributed-system-analysis/pbench/issues/2861