Expand and generalize metadata namespace
PBENCH-448

This allows API clients to access the entire "dataset" and "server"
key namespaces just like "user" and "dashboard"; e.g., to ask for
"dataset" to get all the dataset SQL columns as a JSON document.

This also paves the way to adding a full representation of the agent's
metadata.log metadata without requiring a specific defined and validated
key path for each item.