Add lost tarball state transition images
Nick Dokos has posted state transition diagrams to the old "gh-pages"
web site, but they were lost when that old site was reworked.  This
commit restores those images so that the references to those images in
issue #1315 [1] are restored.

The original commits for these images were 3211b1817 [2] and 9c6bd7a91
[3].

[1] https://github.com/distributed-system-analysis/pbench/issues/1315
[2] https://github.com/distributed-system-analysis/pbench/commit/3211b18174e007534c09b1462aeb62d0a859d58d
[3] https://github.com/distributed-system-analysis/pbench/commit/9c6bd7a918d5d6cfa604ff001fe99fbb0fe87838