Stop using deprecated setuptools method of install
The new way to install our Python3 interfaces is:

    /usr/bin/python3 -m pip install --prefix=${dir} -e .

See https://docs.python.org/3.6/installing/index.html#installing-index.

The new interface allows for the `agent/Makefile` and `server/Makefile`
to create the python CLI interfaces on environments running Python 3.10
and later.

Why do we use `/usr/bin/python3`?  If you invoke pip inside a
virtualenv to install a package, it automatically uninstalls the
previous one from where it was installed, and installs it again in the
`--prefix` path. That is wrong on so many levels. So, to work around
this terrible behavior we have to invoke the non-virtualenv
`/usr/bin/python3` so that the installation works without touching the
virtualenv.