Correct CI / unit tests to work with python 3.10
This also corrects the version of `black` used in the commit hook to
work with Python 3.10.

The version of pytest needs to be at least 6.2.5 in order to work with
Python 3.10.  As a result of this change, we need to correct one test
with its use of `ExceptiionInfo` objects.

Unfortunately, the "usage" message created by the `argparse` module
changed to use "options:" instead of "optional arguments:". We ensure
that only for the two legacy tests that use a python command the output
emitted by older versions of Python3 look like Python 3.10 and later.