Release Notes for `v0.71.0`
This new file, `RELEASE_NOTES.md`, is being added to the `/doc` sub-
directory so that we can code-review the contents of the release notes
like we would any other code change.

The contents of the file will be used for the contents of the
[Alpha](https://github.com/distributed-system-analysis/pbench/releases/edit/v0.71.0-alpha.0),
[Beta](https://github.com/distributed-system-analysis/pbench/releases/tag/v0.71.0-beta.0), and final releases pages on GitHub.

A note about line-wrapping:

The target location of these changes is the GitHub release notes page
for the `v0.71.0-alpha.0` release.  Unfortunately, the way formatting
is handled in a normal GitHub markdown file and how it is handled in the
release note context is different.

Unwrapping the lines so that one paragraph is one line allows us to post
the contents of the release notes directly into the GitHub release notes
page without having to reformat to make it look nice.