Initial changes to support Python 3.10
* Correct CI / unit tests to work with python 3.10

This also corrects the version of `black` used in the commit hook to
work with Python 3.10.

The version of pytest needs to be at least 6.2.5 in order to work with
Python 3.10.  As a result of this change, we need to correct one test
with its use of `ExceptiionInfo` objects.

Unfortunately, the "usage" message created by the `argparse` module
changed to use "options:" instead of "optional arguments:". We ensure
that only for the two legacy tests that use a python command the output
emitted by older versions of Python3 look like Python 3.10 and later.

* Stop using deprecated setuptools method of install

The new way to install our Python3 interfaces is:

    /usr/bin/python3 -m pip install --prefix=${dir} -e .

See https://docs.python.org/3.6/installing/index.html#installing-index.

The new interface allows for the `agent/Makefile` and `server/Makefile`
to create the python CLI interfaces on environments running Python 3.10
and later.

Why do we use `/usr/bin/python3`?  If you invoke pip inside a
virtualenv to install a package, it automatically uninstalls the
previous one from where it was installed, and installs it again in the
`--prefix` path. That is wrong on so many levels. So, to work around
this terrible behavior we have to invoke the non-virtualenv
`/usr/bin/python3` so that the installation works without touching the
virtualenv.