Restore default tool set and deprecate
We are deprecating the notion of the "default" tool set going forward.

Any direct, or indirect, use of the "default" tool set will cause a
warning to be issued by `pbench-register-tool-set` stating that the
"default" tool set will be removed in the Pbench Agent v1.0 release,
and that a named tool set should be used instead.

Further, we are also deprecating the `--toolset` option, and making
the tool set an argument (currently optional as well), since in the
future we'll also require a named tool set be provided.

Note that the "default" tool set is no longer mentioned in any help
output.