Expose tarball metadata.log values through API (#2878)
* Expose tarball metadata.log values through API

PBENCH-682

The filetree code was already parsing metadata.log in order to obtain
the "created" timestamp. This PR formalizes that a bit, converting the
entire config hierarchy (sections/settings) as a JSONOBJECT, which is
persisted in SQL as a "metalog" key value for the dataset.

The "metalog" key value is exposed as a key in the "dataset" namespace,
so that `/datasets/list?metadata=dataset` will return it as part of the
JSON representation of the primary Dataset SQL row. As with any other
metadata namespace, segments can be addressed directly, e.g.,
`dataset.metalog.run.controller`.

NOTE: I'm not deeply in love with the name "metalog"; however I didn't
want to use the obvious "metadata" as it'll all appear nested inside the
outer "metadata" key. I'm open to alternate suggestions!