Switch to tarball MD5 as Dataset resource_id
PBENCH-744

Change the Dataset table to make resource_id the primary key. This must
be unique due to Elasticsearch constraints, and the change allows us to
"demote" dataset name (which is just script+config+date) to metdata that
we can allow the user to set arbitrarily.

This change ripples through the code, resulting in a lot of diffs. Most
of these are fairly straightforward.

Note that this removes legacy test-5.2; PR #2611 also removes this as it
finally drops pbench-server-prep-shim-002 entirely, but the fairly
twisted things test-5.2 does with duplicate and MD5 errors simply don't
apply to the 1.0 server and it's not worth trying to fix it.