Prototype API documentation
PBENCH-754

Begin building Pbench server API documentation in GitHub Markdown,
beginning with some overview material and a few detailed (maybe not
complete) V1 API examples.

We can discuss and refine the limited examples in this PR to get the ball
rolling, and continue to build out APIs in subsequent PRs.