Work around for rpmlint incompatibilities
Starting with `rpmlint` 2.0.0, the `--file` option was changed to
`--rpmlintrc`.  The first encounter with this version of `rpmlint` came
with Fedora 36, but will likely show up soon in other distributions.

To work around this change, we take advantage of two behaviors of
`rpmlint`, one from the pre-2.0.0 version and one from the new version.

For pre-2.0.0, we set the `XDG_CONFIG_HOME` to `${PBENCH_TOP}/utils` so
that our `rpmlint` file containing the configuration we want is found.

For 2.0.0 and later, we copy our `rpmlint` file to a
`pbench-common.rpmlintrc` along side the build location of the .spec
file.