Fix pbench-register-tool-set argument handling (#2904)
Fixes #2903

Invoking pbench-register-tool-set with a named toolset argument, e.g.

   pbench-register-tool-set light

ignores the named argument, and substitutes the default tool set,
resulting in a deprecation warning and registration of the default
tool set.

The option-handling code does one extra shift of the argument list
falling off the end. The fix just eliminates the extra shift.

Also add a simple unit test for this case.