Ensure `fio` arguments are used with `bash -c`
The execution of a fio "client" was failing on the local host because
"bash -c ..." receives multiple, separate arguments.  Since `bash -c`
only considers its first argument, this results in an execution of only
`fio` without the remaining arguments.

The solution is to create a `fio-client` driver script which accepts two
arguments, the directory to execute `fio` from and the port to use with
`--server`.  The invocation of `bash -c` can be properly quoted in the
new context.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>