Improve endpoints API for multiple URI params (#2909)
* Improve endpoints API for multiple URI params

PBENCH-764

I always wanted to expand `/api/v1/endpoints` into a server API schema,
and with the introduction of multiple URI parameters in the inventory
API, I've decided to make an attempt.

Previously I stripped all Flask URI parameter tags from the URI, leaving
a "/" character, and substituting "_" for "/" in the API name. A trailing
"/" works for a single parameter, but "//" isn't very useful for two, nor
is "datasets_inventory_" a great name.

The familiar "api" dict values now give only a single trailing "/"; but
I'm also adding a new "uri" dict which ties the API names to a "schema"
which defines a substitute-able template URI and also provides
parameter definitions with name and type. For example,

```
"datasets_inventory": {
    "template": "http://host/api/v1/datasets/inventory/{dataset}/{path}",
    "params": [
        {"name": "dataset", "type": "string"},
        {"name": "path", "type": "path"}
    ]
}
```