Allow changing the dataset resource name and deletion date
PBENCH-749

Having changed the API to use the unique tarball MD5 as the formal
resource ID, we're free to allow users to select more meaningfull resource
names for filtering and display.

This PR adds a layer allowing the owner of a dataset to use the metadata
PUT API to alter the dataset.name and server.deletion metadata keys with
basic validity checking.