Correct dependencies in agent spec file
This effectively reverts PR #2893.

Simplify the python dependencies as much as possible:

- RHEL7, RHEL9, CentOS-Stream-9 and the Fedoras have a `python3` package
  available from the standard distro repos. The actual version varies by
  distro: 3.6 for RHEL7, 3.9.10 for RHEL9/CentOS-Stream-9, 3.10.0-1 for
  Fedora 35, 3.10.4-1 for Fedora 36.

- RHEL8 and CentOS-Stream-8 have `python36`, `python38` and `python39`
  packages available through the module system.

We use the generic `python3` package on most distros, and standardize on
`python39` for RHEL8/CentOS-Stream-8. That is mostly treated using a
jinja2 loop and a (new) `target.json` file that generates the separate
conditionals for RHEL/CentOS Stream 8 on one hand and the rest on the
other.

Unfortunately, I had to treat RHEL7 separately from the rest because it
is missing a python `cffi` package that all the others have. I haven't
found a clean way to deal with that in `jinja2`, so it is an explicit
RPM conditional in the spec file.

Also, the Fedoras are treated separately, at least for now.

I am testing by setting the release in subscription-manager now
(wherever possible: there is no 7.9 release which I don't understand -
this needs investigation - and there is no 9.1 release yet which is
understandable).

However, that means that RHEL 8.4 and 8.5 have an `ansible` package and
8.6 has an `ansible-core` package available. I've hard-coded
`ansible-core` in the spec file (except on RHEL7) which breaks the
installation on 8.4 and 8.5. I don't know how to deal with this problem.

Modify the base docker file to enable `python39`.  Since we standardized
on `python39` for CentOS-Stream-8, we enable that module and stop
worrying about `python36`.