Add empty targets.json file to server/rpm directory
PR #2901/#2902 added a targets.json file for the agent build
but the change broke the server build. That's because the
change occurred to utils/rpm.mk which is shared by the
agent and server builds.

This fix adds an empty targets.json file to server/rpm. That's
enough to satisfy the jinja2 invocation.