Configure the "python" token used in package names based on version (#2925)
Use a conditional symbol instead of an external file for driving RPM Python dependencies.

PR #2901/#2902 added a targets.json file for the agent build
but the change broke the server build. That's because the
change occurred to utils/rpm.mk which is shared by the
agent and server builds.  This change removes the need for
the targets.json file(s).


Co-authored-by: Nick Dokos <ndokos@redhat.com>