Implement ClientCount check
This should replace the ClientHostAggregteResultCheck and
the last check of SosreportCheck where it checks if the
sos reports have different hostnames. invalid run ids are those
that have more than 2 measurement_idx values for any measurement_title.
It also checks for run ids not in any result data (ie with no corresponding
results and thus invalid)