Use the "epel" chroot for "Rocky Linux"
Backport of PR #2948 (7105b2520).

Set up a map with the `ansible_distribution' value  as key and the
COPR chroot name as value. Use the map to calculate the final
component of the baseurl field in the repo file (except for CentOS 7
which is reporting "CentOS" for the ansible_distribution value but it
requires the "epel" chroot).

The map can be extended easily for other distros ( RHEL clones or
otherwise).