Minor cleanup...
I added a test for an "impossible" condition, given the way Flask handles URI
routing... and for completeness I added a true unit test of the _put_key
method to test it. Shape of things to come?

I may still try another commit on this PR to mock some DB tests to reduce the
reliance on sqlite3; TBD whether I push that commit.