Another try
This encompasses review feedback and also makes an attempt at replacing the
SQLAlchemy engine level (sqlite3) with mocks capable of recording the state
for testing. (E.g., that rollback was called.) The idea is that this provides
the basis of an infrastructure that can be generalized for the other DB test
modules, although this will require additional work.