Pbench Server client library and functional tests (#2941)
* Pbench Server client library and functional tests

PBENCH-810

This adds the beginnings of a simple Pbench Server python client
library (with a `PbenchServerClient` class) and a few initial functional
test modules.

The modules are written in Python using the Pytest framework to
give us convenient test modularization and fixtures. We use a
fixture to establish a Pbench Server connection, for example.

The functional test modules currently include a simple validation
that we can connect to a server and retrieve the server's endpoints,
and a test of Pbench Server user operations including register,
login, logout, profile access/update, and delete.

The `PBENCH_SERVER` environment variable provides the server
host for connection; the `pbench_server_client` fixture will cause failure
if the environment variable isn't defined, or if connection fails.