Drop the UNPACKING/UNPACKED states from dataset
PBENCH-842

With the impending arrival of the cache manager, we're moving away from a
static "unpack" phase. This is a bit of prep work towards that day, removing
the dataset SQL table "UNPACKING" and "UNPACKED" states.

This also allows "re-unpacking" after unpacked data has been aged out (another
problem we'll no longer have after the cache manager) because the unpack tool
is no longer constrained by the state transition rules.

Although this is only tangentially related, I also tried an experiment to make
the end-of-life states more useful (and used) by renaming EXPIRING/EXPIRED to
DELETING/DELETED, using them in the `/datasets/delete` API to put the dataset
into DELETING when we start the operation. (We don't actually use DELETED as
the dataset is completely deleted.)