Fix port requirements in pbench-uperf doc
PBENCH-868

pbench-uperf starts with a server port of 20010, not 20000, so we
fix the doc to agree with the implementation.

The doc does not mention that the local ports on each system need to
be open for uperf to operate successfully. That information plus
information on how to find the range on a system and how to use
`firewall-cmd' to open the range are added.