Disable legacy tests
PBENCH-873

This primarily modifies exec-unittests to stop running the legacy test suites
entirely. Maintaining them is an ongoing and increasingly awkward task as our
server architecture moves away from filesystem-based state.

Interestingly, when I ran a `tox` run, I started getting test failures in the
agent `test_utils.py` because a `BaseServer.kill` call was getting the wrong
computed result code. This appears to be because my system actually had a
process with pid 12345, while that section of the test case runs without a
mock for `os.kill`. The failing test is redundant (as well as wrong) because
there's a comprehensive set of mocked test cases further down. I removed it.