Correct Pbench Server RPM builds
PBENCH-898

With PR #2414, the addition of the put shim server, new python modules
were added, but the tar ball creation process for the source RPM build
was picking up the test subdirectory and the `__pycache__` directory as
well.

We also fixed the `configtools` directory creation to avoid copying the
`__pycache__` directory.