Surrender to the latest version of `black`
I am not sure what exactly happened, but it does not appear possible
to get the old version of `black` ('19') to work in a development
environment.  Further, `tox` seems to have been updated so that it
doesn't work either.  So rather than try to spend a lot more time trying
to find the minimal set of changes, I just gave up and moved to the
the same version of `black` and `flake8` we use on `main`, and updated
and changed whatever was needed to get `tox` to work.