Standardize Python import ordering (#2986)
PBENCH-893

The ordering of Python import statements is a frequent topic of discussion
during PR reviews. It's often not easy when writing code to determine whether
a given module is "standard library" or not, and it's easy to accidentally add
a new symbol out of order.

During a recent review, the `isort` utility was mentioned. I gave it a try but
had some difficulty configuring it to be compatible with our black (despite
the built-in "black" profile), and gave up.

As an excuse to put off dealing with the necessity of breaking my development
VM (and Varshini's dashboard testing), I decided to experiment a bit more to
see if we can solve this issue "once and for all" as we've done with source
format style in general by adopting `black`.

I've now got a configuration (defined in our pyproject.toml file) that causes
`isort` to generate output that makes `black` happy. (This was not trivial,
sadly.) I've also replicated the `black` exclusion list, as that seemed to be
appropriate.

I've added `isort` along with `black` and `flake8` to the lint tests in our
`build.sh` script.

NOTE: see https://pycqa.github.io/isort/docs/configuration/git_hook.html for
instructions on adding `isort` to your git precommit hook. I've also been able
to integrate `isort` into my `vscode` configuration.

`isort` changes 132 files in the tree, which are included in this PR. This
includes obsolete legacy test files, which I decided weren't worth the effort
of excluding.