Correct Pbench Server RPM builds
PBENCH-898

* Correct Pbench Server RPM builds

With PR #2414, the addition of the put shim server, new python modules
were added, but the tar ball creation process for the source RPM build
was picking up the test subdirectory and the `__pycache__` directory as
well.

We also fixed the `configtools` directory creation to avoid copying the
`__pycache__` directory.

----

 * Surrender to the latest version of `black`

I am not sure what exactly happened, but it does not appear possible
to get the old version of `black` ('19') to work in a development
environment.  Further, `tox` seems to have been updated so that it
doesn't work either.  So rather than try to spend a lot more time trying
to find the minimal set of changes, I just gave up and moved to the
the same version of `black` and `flake8` we use on `main`, and updated
and changed whatever was needed to get `tox` to work.

----

Other items addressed:

 * Remove `rpmlint` warnings
 * Bump server version to `0.69.10`
 * Remove wayward `contrib/VERSION`
 * Add use of `rpmlint` when building agent RPM