Unit tests and cleanup
Add a unit test for indexing_tarballs.py. This is somewhat superficial at this
stage, but it establishes a number of mocks necessary to remove the indexer.py
from the loop.

This is not complete, and still hasn't been tested on a real server. Reviewers
beware as much is likely to change and I'll probably squash commits while I
clean the eventual result for review!