`ansible-galaxy` is too "smart" for its own good
The collection build process was taking in the entire directory
hierarchy as part of the collection even if files and directories are
not part of the documented structure.

To fix this, we move all the actual galaxy collection components into
their own "collection" directory, and keep the supporting cast separate.
This has the benefit of not requiring ignore files.

Further, we move the generated artifacts out of the source tree and into
a temporary directory.  The collection path needs to be configured to
use the temporary directory, so this is unfortunately hard-coded to use
`/var/tmp/agent-ansible`.  If really important we can address this
later.