Consolidate ansible directory hierarchy
We remove the extra level of `pbench/agent` under `agent/ansible`, while
moving the playbooks and inventory into their own directory.

We also take the opportunity to remove support for collecting `ara`
system information.  While removing references to it, we also corrected
the references to `pbench-sysinfo-dump` in the documentation tree.

As a result of this refactoring, additional ansible & yaml lint issues
we identified and corrected.

A note about `ansible-galaxy` being too "smart" for its own good:

The collection build process was taking in the entire directory
hierarchy as part of the collection even if files and directories are
not part of the documented structure.

To fix this, we move all the actual galaxy collection components into
their own "collection" directory, and keep the supporting cast separate.
This has the benefit of not requiring ignore files.

Further, we move the generated artifacts out of the source tree and into
a temporary directory.  The collection path needs to be configured to
use the temporary directory, so this is unfortunately hard-coded to use
`/var/tmp/agent-ansible`.  If really important we can address this
later.