Correct Read The Docs builds
Starting around August 22nd, 2022, the Read The Docs builds began
failing, likely due to the obsolescence of the `sphinx-panels` module.

While verifying the build was fixed, we corrected a number of problems
found with `Makefile` targets, the formatting of the configuration file,
two warnings in the user guide, and dropped the Travis-CI documentation
entirely.