Add note on Fedora 36 and its python 3.10 problem resolution
PBENCH-676

Fixes #2884

Fedora has released an update to the python3 package that
fixes the problem. The fix is found in the 3.10.7 version of the
package. After updating, the pbench-agent RPM can be installed
and various commands (e.g. pbench-user-benchmark) that complained
or aborted before, now complete normally.

Details of the fix can be found at https://github.com/fedora-python/cpython/pull/48