Correct UID template resolution
When a template key maps to a non-string value, a error is encountered
like:

    "TypeError: decoding to str: need a bytes-like object, float found"

The code how converts all non-string values to strings, and ignores any
keys which don't map to a value with one of the types: `string`, `int`,
or `float`.

A set of unit tests for this operation is provided, an the function name
was renamed to ensure its intended use is clear.