Correct UID template resolution
Backport of PRs #3029 (665132a89) and #3035 (4822f75af).

When a template key maps to a non-string value, an error is encountered
like:

    "TypeError: decoding to str: need a bytes-like object, float found"

The code now converts all non-string values to strings, and ignores any
keys which don't map to a value with one of the types: `string`, `int`,
or `float`.

A set of unit tests for this operation is provided, and the function
name was renamed to ensure its intended use is clear.

Unfortunately, as the bits have aged the backport of the unit tests is
a bit more difficult.  To make this all work we had to:

 * Remove fallback `elasticsearch` import, it just confused things

 * Move `pytest` invocation under `jenkins-unittests` since we can't
   use `--current-env` with `py3-server`

 * Fix `PYTHONPATH` for unit tests to include the server code

 * Delineate the CI steps a bit more cleanly