Use the same Fedora 35 base image for CI and Dev
To do that, we must remove RPMs which are no longer provided by Fedora
35 and later, `python3-pytest-helpers-namespace` and
`python3-tox-current-env`.  The `pytest-helpers-namespace` module is
still available from PyPI so will be satisfied by the `pip3` install
for the test environment.  The `--current-env` switch is no longer used
with `tox` in our testing environment so that package is not needed at
all.

The `libpq-devel` RPM dependency fails when building the UBI image when
one has multiple PostgreSQL RPM versions.  Rather than trying to solve
that problem, it turns out we don't need to have that RPM installed in
order leverage the `python3-psycopg2` RPM.

In additional we also:

 * Remove unused CI related script

 * Drop Fedora 34 from RPM builds

 * And correct some long comment lines.