Rework CLI commands to use enhanced `ToolGroup`
A significant re-write to the `ToolGroup` is taken to provide APIs for
on-disk operations entirely encapsulated by that class.  All existing
commands now use the `ToolGroup` class to accomplish their operations.

The `ToolGroup` class uses the `filelock` Python3 module
(`python3-filelock` RPM) to ensure that on-disk state is not mutated
while it is being read or written.

Any changes to the on-disk state require an existing `ToolGroup` object
to reload its in-memory state from disk when referenced.