Correct version of Python used by server to 3.9
While inspecting the current production deployment we have of the
v0.69.* Pbench Server we found that it was using Python 3.6 instead of
the intended 3.8.

However, the Jenkins unit test environment will execute the tests
under Python 3.9 since the support for running under Python 3.6 is
precluded by the Fedora 33 environment used.

We fix the version of Python for the Pbench Server by moving to Python
3.9, using `%py3_shebang_fix` to explicitly use `/usr/bin/python3.9`
in the command line interface headers, and using `pip3.9` to do the
local module installs.

With an additional correction to the `tox` environments for the Pbench
Server, `server` and `py3-server`, all server tests are executed under
Python 3.9, since the default for Fedora 33 (the CI container in use) is
also Python 3.9.

The minimum version of Python required for the Pbench Agent is 3.6,
which comes from the fact we support RHEL 7.  To that end, we make a
change to force the `tox -e agent` environment to run under 3.6.

But, the Pbench Agent tests under the CI will still use the default
version of Python 3 provided by Fedora 33, which is Python 3.9.